requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
click==8.1.7
    # via -r requirements.in
colorama==0.4.6
    # via click
pydantic==2.8.2
    # via -r requirements.in
pydantic-core==2.20.1
    # via pydantic
pyyaml==6.0.1
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   pydantic
    #   pydantic-core
